src/test/java/mx/iteso/ut/TortillaMaizTest.java
package mx.iteso.ut;

import org.junit.Test;
import org.junit.Before;
import static org.junit.Assert.*;

public class TortillaMaizTest {
	TortillaMaiz tortillaMaiz;
	
	@Before
    public void setUp() {
		tortillaMaiz = new tortillaMaiz();
    }

    @Test
    public void testCurrentTemperature() {
    	tortillaMaiz.setCurrentTemperature(10);
        assertEquals(10, tortillaMaiz.getCurrentTemperature());
    }

    @Test
    public void testFalseToast() {
    	tortillaMaiz.toast(false);
        assertFalse(tortillaMaiz.isToasted());
    }

    @Test
    public void testTrueToast() {
    	tortillaMaiz.toast(true);
        assertTrue(tortillaMaiz.isToasted());
    }

    @Test
    public void testToasting() {
        assertEquals(20, tortillaMaiz.getToastTemperature());
    }


}
